Avoid pulling libdrm.so from the AOSP host build
When binaries depending on virglrenderer on the host are linked to it,
it will cause libdrm.so to be taken from the host build done by the
Android build system (i.e. out/host/linux-x86/lib64/libdrm.so), but
usually this library is used with binary GPU drivers that are built
out-of-tree, and they will use the libdrm.so from the host system
(i.e. /usr/lib/x86_64-linux-gnu/libdrm.so.2).

This workaround is only applied to the host "libvirglrender.so" target,
which is only used by crosvm currently.

Change-Id: I84d7227bb6fb0118bd05034c1633f6604d8d838f